Bump cloud.google.com/go to get timeout fix
The commit we're interested in is
googleapis/google-cloud-go@fbf2f51 ,
which disables an aggressive 2-second timeout for awaiting headers from
the compute metadata service.

When workload identity is enabled on a cluster, a different endpoint is
used under the hood for compute metadata, and this endpoint can often
take more than 2 seconds to return headers.

Signed-off-by: Joshua Kwan <joshua.kwan@liveramp.com>